introduce new module prover::serialization
