+++ prover/src/serialization.rs
@@ -1,0 +1,120 @@
+use plonky2::{
+    gadgets::{
+        arithmetic::EqualityGenerator,
+        arithmetic_extension::QuotientGeneratorExtension,
+        range_check::LowHighGenerator,
+        split_base::BaseSumGenerator,
+        split_join::{SplitGenerator, WireSplitGenerator},
+    },
+    gates::{
+        arithmetic_base::{ArithmeticBaseGenerator, ArithmeticGate},
+        arithmetic_extension::{ArithmeticExtensionGate, ArithmeticExtensionGenerator},
+        base_sum::{BaseSplitGenerator, BaseSumGate},
+        constant::ConstantGate,
+        coset_interpolation::{CosetInterpolationGate, InterpolationGenerator},
+        exponentiation::{ExponentiationGate, ExponentiationGenerator},
+        lookup::{LookupGate, LookupGenerator},
+        lookup_table::{LookupTableGate, LookupTableGenerator},
+        multiplication_extension::{MulExtensionGate, MulExtensionGenerator},
+        noop::NoopGate,
+        poseidon::{PoseidonGate, PoseidonGenerator},
+        poseidon_mds::{PoseidonMdsGate, PoseidonMdsGenerator},
+        public_input::PublicInputGate,
+        random_access::{RandomAccessGate, RandomAccessGenerator},
+        reducing::{ReducingGate, ReducingGenerator},
+        reducing_extension::{
+            ReducingExtensionGate, ReducingGenerator as ReducingExtensionGenerator,
+        },
+    },
+    get_gate_tag_impl, get_generator_tag_impl,
+    hash::hash_types::RichField,
+    impl_gate_serializer, impl_generator_serializer,
+    iop::generator::{
+        ConstantGenerator, CopyGenerator, NonzeroTestGenerator, RandomValueGenerator,
+    },
+    plonk::config::{AlgebraicHasher, GenericConfig},
+    read_gate_impl, read_generator_impl,
+    recursion::dummy_circuit::DummyProofGenerator,
+    util::serialization::WitnessGeneratorSerializer,
+};
+use plonky2_field::extension::Extendable;
+use plonky2_u32::gates::{
+    add_many_u32::{U32AddManyGate, U32AddManyGenerator},
+    arithmetic_u32::{U32ArithmeticGate, U32ArithmeticGenerator},
+};
+use std::marker::PhantomData;
+
+/// The same as `plonky2::util::serialization::generator_serialization::default::DefaultGeneratorSerializer`
+/// but with added `U32AddManyGenerator` and `U32ArithmeticGenerator`.
+#[derive(Debug, Default)]
+pub struct GeneratorSerializer<C: GenericConfig<D>, const D: usize> {
+    pub _phantom: PhantomData<C>,
+}
+
+impl<F, C, const D: usize> WitnessGeneratorSerializer<F, D> for GeneratorSerializer<C, D>
+where
+    F: RichField + Extendable<D>,
+    C: GenericConfig<D, F = F> + 'static,
+    C::Hasher: AlgebraicHasher<F>,
+{
+    impl_generator_serializer! {
+        GeneratorSerializer,
+        ArithmeticBaseGenerator<F, D>,
+        ArithmeticExtensionGenerator<F, D>,
+        BaseSplitGenerator<2>,
+        BaseSumGenerator<2>,
+        ConstantGenerator<F>,
+        CopyGenerator,
+        DummyProofGenerator<F, C, D>,
+        EqualityGenerator,
+        ExponentiationGenerator<F, D>,
+        InterpolationGenerator<F, D>,
+        LookupGenerator,
+        LookupTableGenerator,
+        LowHighGenerator,
+        MulExtensionGenerator<F, D>,
+        NonzeroTestGenerator,
+        PoseidonGenerator<F, D>,
+        PoseidonMdsGenerator<D>,
+        QuotientGeneratorExtension<D>,
+        RandomAccessGenerator<F, D>,
+        RandomValueGenerator,
+        ReducingGenerator<D>,
+        ReducingExtensionGenerator<D>,
+        SplitGenerator,
+        WireSplitGenerator,
+        U32AddManyGenerator<F, D>,
+        U32ArithmeticGenerator<F, D>
+    }
+}
+
+/// The same as `plonky2::util::serialization::gate_serialization::default::DefaultGateSerializer`
+/// but with added `U32AddManyGate` and `U32ArithmeticGate`.
+#[derive(Debug, Default)]
+pub struct GateSerializer;
+
+impl<F: RichField + Extendable<D>, const D: usize>
+    plonky2::util::serialization::GateSerializer<F, D> for GateSerializer
+{
+    impl_gate_serializer! {
+        GateSerializer,
+        ArithmeticGate,
+        ArithmeticExtensionGate<D>,
+        BaseSumGate<2>,
+        ConstantGate,
+        CosetInterpolationGate<F, D>,
+        ExponentiationGate<F, D>,
+        LookupGate,
+        LookupTableGate,
+        MulExtensionGate<D>,
+        NoopGate,
+        PoseidonMdsGate<F, D>,
+        PoseidonGate<F, D>,
+        PublicInputGate,
+        RandomAccessGate<F, D>,
+        ReducingExtensionGate<D>,
+        ReducingGate<D>,
+        U32AddManyGate<F, D>,
+        U32ArithmeticGate<F, D>
+    }
+}
